Makefile: Do not generate jekyll-gh-pages Docker image if it exists
In order to save build time, do not generated the `jekyll-gh-pages`
Docker image if it already exists. The `jekyll-gh-pages` Docker image is
the generic based image used by the current deployment. Building it
takes some time. Also, once built, it doesn't need to be rebuit, unless
some critical security actions need to be undertaken.

Update the `build` rule in the `Makefile` to check if the
`jekyll-gh-pages` Docker image already exists. If it doesn't exist,
create it. Otherwise, move to building the Docker image for the
deployment (`jekyll-gh-pages-site`).

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@upb.ro>